api/requirements.txt
fastapi
uvicorn

SQLAlchemy
pydantic
psycopg2
email-validator
python-multipart
python-jose
passlib
bcrypt
requests
spacy